:truck: api: move get-shareable-certificate usecase from lib to src
