+++ api/src/certification/results/domain/usecases/get-shareable-certificate.js
